+++ packages/dust-wallet/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @midnight-ntwrk/wallet-sdk-dust-wallet
 
+## 1.0.0-beta.14
+
+### Patch Changes
+
+- 8b8d708: chore: update ledger to version 7.0.0-rc.1
+- Updated dependencies [8b8d708]
+  - @midnight-ntwrk/wallet-sdk-shielded@1.0.0-beta.15
+  - @midnight-ntwrk/wallet-sdk-address-format@3.0.0-beta.11
+  - @midnight-ntwrk/wallet-sdk-prover-client@1.0.0-beta.13
+  - @midnight-ntwrk/wallet-sdk-capabilities@3.0.0-beta.11
+  - @midnight-ntwrk/wallet-sdk-node-client@1.0.0-beta.12
+  - @midnight-ntwrk/wallet-sdk-utilities@1.0.0-beta.10
+  - @midnight-ntwrk/wallet-sdk-indexer-client@1.0.0-beta.16
+
 ## 1.0.0-beta.13
 
 ### Patch Changes