+++ packages/facade/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @midnight-ntwrk/wallet-sdk-facade
 
+## 1.0.0-beta.15
+
+### Patch Changes
+
+- 8b8d708: chore: update ledger to version 7.0.0-rc.1
+- Updated dependencies [8b8d708]
+  - @midnight-ntwrk/wallet-sdk-unshielded-wallet@1.0.0-beta.17
+  - @midnight-ntwrk/wallet-sdk-shielded@1.0.0-beta.15
+  - @midnight-ntwrk/wallet-sdk-address-format@3.0.0-beta.11
+  - @midnight-ntwrk/wallet-sdk-dust-wallet@1.0.0-beta.14
+
 ## 1.0.0-beta.14
 
 ### Patch Changes