+++ packages/utilities/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @midnight-ntwrk/wallet-sdk-utilities
 
+## 1.0.0-beta.10
+
+### Patch Changes
+
+- 8b8d708: chore: update ledger to version 7.0.0-rc.1
+
 ## 1.0.0-beta.9
 
 ### Patch Changes